+++ docs/components/content/Slider/SliderBasic.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+  import { Slider } from '@lite-space/ui'
+
+  const slederValue = ref(30)
+</script>
+
+<template>
+  <Slider v-model="slederValue" />
+</template>