+++ packages/components/Slider/src/types.d.ts
@@ -1,0 +1,8 @@
+interface LtSliderProps {
+  min: number
+  max: number
+  step: number
+}
+
+export type SliderProps = Partial<LtSliderProps>
+export type SliderInstance = InstanceType<typeof import('./Slider.vue')['default']>