+++ packages/theme/src/shortcuts/Slider.ts
@@ -1,0 +1,7 @@
+export const Slider = {
+  'lt-slider': 'cursor-pointer relative',
+  'lt-slider-track': 'bg-gray-100 rounded-md h-1.5',
+  'lt-slider-track--active': 'bg-black/85 h-full rounded-md',
+  'lt-slider-handle': 'absolute left-0 top-50% -translate-50%',
+  'lt-slider-handle-inner': 'w-4 h-4 rounded-full border border-solid border-gray-300 bg-white'
+}