+++ packages/components/Tooltip/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lite-space/tooltip
 
+## 0.0.3
+
+### Patch Changes
+
+- @lite-space/slider@0.0.0
+  @lite-space/tooltip@0.0.3
+  @lite-space/ui@0.6.0
+  @lite-space/theme@0.6.0
+
 ## 0.0.2
 
 ### Patch Changes